requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
alembic==1.11.1
arrow==1.2.3
binaryornot==0.4.4
blinker==1.6.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
cookiecutter==2.1.1
cycler==0.11.0
databricks-cli==0.17.7
docker==6.1.3
entrypoints==0.4
evidently==0.3.3
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.1
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mlflow==2.4.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.3
pyarrow==12.0.1
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.17
sqlparse==0.4.4
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.1.0
typing-extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.15.0